hermit_/webbacksoftware/requirements.txt
amqp==2.5.0
asn1crypto==0.24.0
betterpath==0.2.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Django==1.11.11
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-js-asset==1.2.2
django-redis==4.10.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
fdfs-client-py==1.2.6
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
kombu==4.6.3
l==0.11.0
MarkupSafe==1.1.1
mutagen==1.42.0
Pillow==6.0.0
Pillow-PIL==0.1.dev0
prettytable==0.7.2
pycparser==2.19
pyecharts==1.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
urllib3==1.25.3
uWSGI==2.0.18
vcversioner==2.16.0.0
vine==1.3.0
Werkzeug==0.15.4
zope.interface==4.6.0
